RDF triples iterator for `enum` types.
